requirements.txt
arcade==3.3.2
attrs==25.3.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.0.0
pycparser==2.23
pygame==2.6.1
pyglet==2.1.8
pymunk==6.9.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytiled_parser==2.2.9
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
wheel==0.45.1
